Backend

+++ server/models/business-management/recipe_detail.model.js
@@ -2,7 +2,7 @@ const { DataTypes } = require('sequelize');
 const sequelize = require('../config.database');
 
 const RecipeDetail = sequelize.define('recipe_detail', {
-    quantity: DataTypes.INTEGER
+    quantity: DataTypes.FLOAT
   },
   {
     timestamps:false,
